revise depth normalization in Starling VQSR -- STARKA-162
Various changes to VQSR depth normalization:
0) normalize GQX and QUAL scores (divide by mean depth)
1) change normalization to use chromosome-specific mean coverage, in order to avoid strange behavior for haploid (male chrX and chrY) and mitochondrial (chrM) regions
2) change normalization to use a floating-point depth mean
3) update logistic model to reflect the latest changes Nondas made in model training code (GQX and QUAL normalization, removing some rank sum statistics from SNP models to improve depth robustness)

+++ src/c++/lib/starling_common/gvcf_locus_info.hh
@@ -305,7 +305,7 @@ struct indel_info
 
     void calc_vqsr_metrics(); //calculate the final VQSR metrics from the indel_key data
 
-    std::map<std::string, double> get_qscore_features(int chrom_depth);
+    std::map<std::string, double> get_qscore_features(double chrom_depth);
 
     pos_t pos;
     indel_key ik;
@@ -368,7 +368,7 @@ struct site_info
         }
     }
 
-    std::map<std::string, double> get_qscore_features(int chrom_depth);
+    std::map<std::string, double> get_qscore_features(double chrom_depth);
 
     bool
     is_het() const
